requirements.txt
:wq
Babel==1.3
Django==1.8.3
django-appconf==1.0.1
django-compressor==1.5
django-cors-headers==1.0.0
django-easy-pdf==0.1.0
django-extra-views==0.6.4
django-haystack==2.3.1
django-localflavor==1.1
django-oscar==1.1
django-oscar-datacash==0.8.3
django-oscar-paypal==0.9.5
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.3
factory-boy==2.4.1
gunicorn==19.3.0
html5lib==0.999999
lesscpy==0.10.2
mock==1.0.1
phonenumbers==6.3.0
Pillow==2.9.0
ply==3.8
purl==1.1
pycountry==1.12
PyPDF2==1.25.1
pytz==2015.4
reportlab==2.7
requests==2.7.0
six==1.9.0
sorl-thumbnail==12.2
stripe==1.23.0
Unidecode==0.4.18
uWSGI==2.0.11.1
waitress==0.8.10
xhtml2pdf==0.0.6
